Update README.md
A full web application with a beautiful user interface based on .NET MVC for hotel reservations with SQL database. U can add new client info such as name, nationality, residence, and so on. there is another page that displays all clients in the hotel with their info. there is a specific section for adding new payment methods and types of rooms whether single or double as u want.